requirements.txt
fastapi==0.111.1
requests==2.31.0
dnspython==2.3.0
asyncio==4.0.0